requirements.txt
Django>=4.2
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
celery[redis]
qrcode
reportlab
boto3
minio
django-cors-headers
gunicorn
Pillow
python-magic
opencv-python-headless
pyzbar
numpy
django-storages
minio
